refactor: simplify page factory
